Loading additional Javascript-files: resources/js/NetspocWeb/listpanel/NetworkResourcesList.js resources/js/NetspocWeb/button/PrintButton.js resources/js/NetspocWeb/cardpanel/PrintActive.js
print.css now in correct directory.


Removed obsolete file dienste.js.working_prototype


Added print-buttons and changed from tab-panels to using card-panel with button-group (in order to be able to have a print-button on the same tbar).


Standard button with a printer icon that calls printView on the listview contained in the parent panel of the button.


Card-panel with feature of printing currently active panel.


Moved to correct directory.