Show service manager; better multiple owners.
Use a button with group icon to toogle between different owners.
TriggerField isn't used any longer.
The button is dynamically shown or hidden, depending on the number of owners.
This requires layoutOnCardChange : true
Service manager is added to list of owners.
But label changes to "Verwalter:" if manager is shown.