Fixed choose-network-bug in netspoc.fcgi
Fixed bug that made owner-services appear in used
services when networks were chosen.
Led to errors like "TypeError: g is undefined".

The list of rules was undefined for those "wrong"
services leading to a JSON-response with a "success"-property
but without a "record"-property. So the root for the JSON-reader
was undefined resulting in the above TypeError.