"Eigene Netze" now works in a basic version.
"Netzauswahl"-feature is still missing, because it will be
moved to a different place in the GUI and it will be
implemented more slim.

Added select0-method to Abstract-base-grid, so all derived
grids can simply call it and have their first row selected
(if there is a row that can be selected).