Fixed bug: workaround for unwanted tree autoload.
Clicking on the arrow of the Diff tree root sends a load
to the tree store, although configured with "autoLoad:false".

When no items was selected in the version combo, the param
"version=none" is sent to the backend.

So now the backend returns an empty list in that case.