NetworkResources is now a grid with grouping.
The list of own networks has become selectable, so now
several networks can be selected, which made the former
display of resources for one selected network useless.

Now the network names are displayed as header and the
resources of that network are grouped below.

TODO: make the choice of networks actually active and
      change "Eigene Netze" button to reflect change
      of data basis.