Initial checkin on the way to port to ExtJs 4
Removed old Javascript-code in htdocs/resources.

Added minimal initial skeleton of an ExtJs 4 App (adhering
to architecture guide from Sencha).

Modified netspoc.fcgi with a routine return_data that returns some
data as a list, so that a minimal version showing data can
be established before porting the NetspocStore and Appstate
infrastructure to the new version.

As soon as the basic infrastructure is ported to the new
version of the ExtJs-framework, the application will be ported
tab by tab:
"Dienste, Freischaltungen", then
"Eigene Netze", followed by
"Diff", and finally
"Berechtigungen".