Improved network selection in "Eigene Netze".
Selecting all networks did not scale well.
The selection itself not and the display of contained
resources even less.
So we got rid of "select all"-feature altogether for which
there never was a convincing use case anyway.

Canceling a selection made is a valid use case, however,
so there is now a button next to the confirm button that
does exactly that.
Pressing it deselects all networks, empties the resources
display panel and fires a click event for the confirm button,
so that the empty selection is activated.