Minor changes in card "Eigene Netze".
Got rid of "Netze"-button, which is not needed anymore, since
the button "Router" next to it was removed.

Shortened button text for confirm and cancel buttons by
prepending a label with common button text "Netzauswahl".

Attribute headerWidth for CheckboxSelModel needs to be
12 for ExtJs 4.2.0 and 24 for 4.2.1 for the checkbox column to look
right. Since we downgrade to 4.2.0 because of a bug in handling tabchange
event in 4.2.1, we need to set the value of this attribute back to 12.