Implemented tooltips for Search Form.
Tooltips can now, thanks to Datatip.js be added
easily to all kinds of components, including forms.
Either a "tooltip"-attribute is used as html
for the Tooltip or a template (provided via tpl attribute)
gets filled with values stored in the "data"-attribute.

Longer tooltips can be loaded via a loader config parameter
(see the textfields in the search window for an example).