Fix bug in search form: own and used services.
Checkboxes for own and used services were not heeded.
Which services were searched was determined by param
"relation" which was not the desired behaviour.

For the store load when search is started this param is set
to an empty string so that the checkboxes are paid attention to
in the backend.