Send search params to URL of user and rules store
If we are in search mode, the params entered there are
now sent to the backend for the User- and the Rules-store,
too.

This way, the results can be filtered to match only resources
found in the search. This behaviour is more consistent with
the way results get filtered when selecting a subset of own networks
in "Eigene Netze".