Fix: make dzil test work
dzil test failed with an error about missing TestRunner plugins.
To amend that, the [MakeMaker] directive was added to "dist.ini".
The downside is that a Makefile.PL is generated and distributed with the tar archive,
too, which can be a bit misleading. A user not familiar with this software might
feel invited to use "make; make test; make install" to install it, which
is not the way it is designed (at least not yet). This software is, so far,
meant to be installed manually by extracting the tar archive (and reloading or
restarting the backend if that part of the code was changed).

[TestRelease] was added to dist.ini, too, so that the code tests are
run before a release is actually done.