Fix issue #1 and add enhancement of issue #2
The rules displayed after a search are no longer filtered
by default and expanded accidentally when checking and
unchecking the "User expandieren" checkbox.

Instead, rules are still filtered by default, but checking
and unchecking does not change that anymore.

It is now possible to choose whether to filter the rules or not
by (un-)checking a checkbox called "Regeln filtern".