Add: collapse network ressources if too many
A threshold of 100 is defined. If the resources contained in a network
are more than that threshold, the grid group is initially
displayed collapsed. This speeds up the display of large networks
and is good, if you only want to select certain networks and
don't care about the contained resources.

This closes #10.