Add: ability to filter users according to search
When in search mode, the resulting rules can be filtered to display
only those that have objects as source or destination that match the search
parameters.

Now this is possible for the Users, too, when the appropriate tab is
selected and the filter-checkbox is activated. If it is deactivated, then
all objects contained in "User" are displayed.