Fix: IE problems with rendering many rows.
For owners with many (thousands of) services, IE became unbearably
slow when trying to render those rows in a grid. Firefox responds
in 2s when it takes IE to 3,5 minutes to render 800 rows.

To fix this, there are two solutions:
1. A buffered store and a paging grid in the frontend
   and buffering implemented in the backend.
2. A buffered renderer (native plugin from Sencha).

Solution nr. 2 has been implemented for all grids inheriting from
"Abstract" grid.