Add: email-workflow, first step: add user-object
A simple form of a workflow will be implemented as a first step
on the way of enabling Policy-Web to actually be used to change
data in Netspoc.

In this first step, Policy-Web will not change Netspoc directly
but generate emails containing tasks that get sent to the user.
After being checked by the user the task at hand can be sent to
communication management for further processing.

This commit contains changes for generating a task email for
adding a new object to the "User" part of the currently selected
service.