Add: Window to delete object from "User" task mail
Implements a window with a form that enables the user to select an object from existing
"User"-objects and generate an email with a task to delete
the selected object.

That email has to be and afterwards sent to communication management to gather necessary
permissions and generate a task for network operations to really change data in Netspoc.