Only display change rule buttons for admins
The buttons to change rules should be only displayed if the user
logged in for current owner is an admin for that owner, not for watchers.

This commit also takes back changes made earlier that disabled the buttons
to change "User"-data if there was only one item in the list of users.
Heinz Knutzen correctly argued that it is perfectly fine to delete the
single item present, because there can be more items relevant for that
service, which are just not visible for the current user.