Now use "vis.js" instead of D3
Couldn't get D3 to play nice with ExtJs using the force layout.
Now "vis.js" gets used to display graph.

This is a commit before beginning to implement changes due to
presentation of prototype to Christof Rode.

Will get rid of combo box with own resources in favour of
a list of own resources with checkboxes, very similar to
how the own networks and the display of their resources is
implemented.