Add "Mandant" and more IP consistency checks
A combo-box containing selectable business units ("Mandanten") is added
to window that generates mail to change "User" of a service.
Those business units come from a text file. The path to that file is
configured in policyweb.conf (attribute "business_units").

The previous changes for an overview independent of services will be
part of this release, but the users won't notice, because the button
to activate the window is not shown (commented out in source code).

Changes to bin/cleanup_daily:
- Remove users from /home/netspocweb/users if they are no longer known
  as admins or watchers in Netspoc.
  Deletes user-file only if it is older than a week in order to make it possible
  to revert errorneous recent changes to admins or watchers without anybody noticing.
- Remove session files older than a year.