Add lnsh-policy as target for forwarded mails
Task emails should be forwarded to dataportlnsh-policy@dataport.de
after evaluation.

That email-address was missing in zthe templates.

Fix: test search_ip.t
This test lacked two config options:
 "about_info_template"  : "",
 "business_units"       : "",
and it still tried to find export.pl in the bin-directory of this
repository. export.pl was moved to Netspoc, so the path to that binary
had to be changed.