Add: test for compatibility mode in IE
No one seems to see the warning on the front page of PolicyWeb to
deactivate compatibility mode in IE for the site to work properly.

So now a warning message pops up in a window on startup if compat mode
is detected. To get rid of that window, you have to activate a checkbox
to confirm that you have taken note of the warning. Checking that box
will deactivate the warning for future application starts (stored
is session).