Fix: use name of input field instead of id
Input text field for email and password had an id so far.
Heinz Knutzen removed those attributes when adding ldap login.

This made the login test fail, because those fields
were searched for by using the id.

Now the name is used instead of id and the lines giving the
focus to email input field use the name, too.