Remove owner selection window. Resize service list
When there is no session or session is invalid (e.g. expired)
there used to be an owner selection window. This is no longer
needed, because the best owner is now determined and selected
automatically.

Instead, the user is now redirected to the login page if there
is no session or an invalid one.

Service list was configured to have a fiexed width (310px) and
to be resizable. Resizing never worked, however, because
ExtJs throws an internal (non-fatal) error when trying to do
so. Because it would be difficult to track this error in the
framework itself, the width now is configured to take up 20%
of the viewport space, which should nowadays be more than
enough, so that there (hopefully) will be no more need for
resizing ...