Remove IP sorting from sub adapt_name_ip_user
The src and dst objects are stored in array refs. The order
is important here, because IPs and names are found at the same
index.
The sorting changes that order and might evend remove elements
by using a hash in the sorting function.
So the sorting is removed here and should be done in the
client (UI).