Fix sorting in rules of a service
Correct sorting depends on the checkbox that switches between
display of object names and IP addresses.

SortType for "Quelle" and "Ziel" is initially set to "asIP",
which is fine for IP addresses displayed in the beginning for
either src or dst. The other side displays "User" to begin
with, which is handled by "asIP" just fine.

When the checkbox is toggled, the SortType of the rules model
fields "src" and "dst" are toggled between "asIP" and
"asUCText".