Aggregate IP addresses in services overview, too
There are many any objects that share the same IP address.
Those are aggregated to one entry if IP addresses are shown,
but all objects are displayed in case of names to be shown
(checkbox).
This behaviour was missing in the overview window of all services.
The function "bold_user" providing this is moved to common.js
in order to have only one place to change in case of bugs or
changes (DRY-principle).