Fix bug introduced with has_user changes
In those changes data was always sent for src and dst,
no matter what expand_rules pointed to.
Now if exapnd_user is false and has_user is "src",
the src-attribute will be [].
If expand_user is false and has_user is "dst",
the dst-attribute will be [].
If expand_user is true, both will be filled with data.