Fix bug in func getServicesAndRules
If list was empty, an uninitialized value was returned.
func writeRecords returned "records:null" in this case,
but it should return "records:[]".
Fixed both. Result gets initialized as empty slice and
writeRecords now handles this case properly, if by
accident there is an uninitilized value passed to it as param.