Fix bug with sub-dir, generate pass hash with Perl
A subdir in the temp-dir needs to be created before files
can be written to it.

The password hash generated by go did not match the one generated
by Perl.
Finding a golang module capable of writing SHA256 hash in the
needed format was not successful, so we use the existing Perl
script add_user_pass to genrate those hashes.