Move creation of tmp dir to individual test cases
Up to now the temporary workdir was set in the wrapper starting
all the .t files as tests.
This led to errors when creating password hashes stored in
user session store file. When calling a second test the user
session file already exists leading to an error in the Perl
script that tries to create a new user session file to store
the hash in.
Now a new workdir is created for each individual test and the
PerlTestServer is started for each test, too.