Migrate /get_diff to go, status is unfinished
The diff for users and for own and used services is already
working.
The diffs for global services and objects is still missing.

policydiff.go has the required code to generate the diffs.
send-diff.go has code for sending diffs via mail and is
used to test the migrated code, too.